+++ src/display_drivers/latex_display_driver.hpp
@@ -1,0 +1,16 @@
+#ifndef MDL_LATEX_DISPLAY_DRIVER_HPP
+#define MDL_LATEX_DISPLAY_DRIVER_HPP
+
+#include "display_driver.hpp"
+
+/**
+ * @brief LaTeX implementation of the displaydriver class
+ **/
+class LaTeXDisplayDriver : public DisplayDriver {
+	public:
+		LaTeXDisplayDriver() {};
+		virtual ~LaTeXDisplayDriver() {};
+		virtual void display(Document *doc);
+};
+
+#endif // MDL_LATEX_DISPLAY_DRIVER_HPP